+++ CHANGELOG.md
@@ -9,7 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Enable drag-and-drop rows via `@dnd-kit` package.
+- option `dragAndDropOptions` to enable drag-and-drop rows via `@dnd-kit` package.  
+
+### Added
+
+- option `cellClassName` added to the column meta to add class names to every cell of a column
+- option `headerClassName` added to the column meta to add class names to every header of a column
+- option `footerStyle` added to the column meta to add styles to every footer of a column
+- option `footerClassName` added to the column meta to add class names to every footer of a column
+
+## [4.1.0] - 2024-03-13
+
+### Added
+
+- option `hidePageSizeChange` added to hide the possibility to change the page size
 
 ## [4.0.0] - 2024-03-06
 
@@ -255,7 +268,8 @@ columnHelper.display({
 
 - created package :tada:
 
-[unreleased]: https://github.com/neolution-ch/react-data-table/compare/4.0.0...HEAD
+[unreleased]: https://github.com/neolution-ch/react-data-table/compare/4.1.0...HEAD
+[4.1.0]: https://github.com/neolution-ch/react-data-table/compare/4.0.0...4.1.0
 [4.0.0]: https://github.com/neolution-ch/react-data-table/compare/3.8.0...4.0.0
 [3.8.0]: https://github.com/neolution-ch/react-data-table/compare/3.7.0...3.8.0
 [3.7.0]: https://github.com/neolution-ch/react-data-table/compare/3.6.0...3.7.0